+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file based on the
 
 ### Improvements
 
+- added support for the "explain" flag of AnalyzeAPI [PR-1254](https://github.com/ruflin/Elastica/pull/1254)
+
 ### Deprecated